+++ test/test_filesytem_queue.rb
@@ -1,12 +1,11 @@
 # frozen_string_literal: true
 
-require "test_helper"
-require "fileutils"
+require 'test_helper'
+require 'fileutils'
 
 class FilesystemQueueTest < Minitest::Test
   def setup
-    @queue_dir = "test/test_queue"
-    FileUtils.rm_rf(@queue_dir)
+    @queue_dir = 'test/fixtures/test_queue'
     @queue = FilesystemQueue::Queue.new(@queue_dir)
   end
 
@@ -14,29 +13,25 @@ class FilesystemQueueTest < Minitest::Test
     FileUtils.rm_rf(@queue_dir)
   end
 
-  def test_that_it_has_a_version_number
-    refute_nil ::FilesystemQueue::VERSION
-  end
-
-  def test_initialize_creates_directories_and_index_file
+  def test_initialize_creates_directories_and_index
     assert Dir.exist?(@queue.instance_variable_get(:@jobs_dir))
     assert Dir.exist?(@queue.instance_variable_get(:@completed_dir))
     assert Dir.exist?(@queue.instance_variable_get(:@failed_dir))
-    assert File.exist?(@queue.instance_variable_get(:@index_file))
+    assert_equal [], @queue.instance_variable_get(:@index)
   end
 
   def test_enqueue_adds_job_to_queue
-    job = { task: "test_task" }
+    job = { task: 'test_task' }
     @queue.enqueue(job)
     job_file = Dir.glob("#{@queue.instance_variable_get(:@jobs_dir)}/*.json").first
     assert File.exist?(job_file)
     assert_equal job.to_json, File.read(job_file)
-    assert_equal job_file, File.read(@queue.instance_variable_get(:@index_file)).strip
+    assert_includes @queue.instance_variable_get(:@index), job_file
   end
 
   def test_dequeue_removes_and_returns_oldest_job
-    job1 = { task: "test_task_1" }
-    job2 = { task: "test_task_2" }
+    job1 = { task: 'test_task_1' }
+    job2 = { task: 'test_task_2' }
     @queue.enqueue(job1)
     @queue.enqueue(job2)
 
@@ -44,12 +39,8 @@ class FilesystemQueueTest < Minitest::Test
     assert_equal job1, dequeued_job
     assert File.exist?(dequeued_job_file)
 
-    remaining_jobs = File.read(@queue.instance_variable_get(:@index_file)).strip.split("\n")
+    remaining_jobs = @queue.instance_variable_get(:@index)
     assert_equal 1, remaining_jobs.size
-    assert_includes remaining_jobs.first, "job_"
-  end
-
-  def test_dequeue_returns_nil_if_queue_is_empty
-    assert_nil @queue.dequeue
+    assert_includes remaining_jobs.first, 'job_'
   end
 end