+++ test/test_helper.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-$LOAD_PATH.unshift File.expand_path("../lib", __dir__)
-require "filesystem_queue"
+$LOAD_PATH.unshift File.expand_path('../lib', __dir__)
+require 'filesystem_queue'
 
-require "minitest/autorun"
+require 'minitest/autorun'