+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+## [0.2.0] - 2024-09-16
+- Replaced old index for an in-memory index for job tracking
+- Added `cleanup` method to delete files and directories created by the queue
 ## [0.1.1] - 2024-09-13
 - Added unit tests
 - Fixed Github Actions